requirements.txt
Flask
language-tool-python
nltk
numpy
pyspellchecker
scikit-learn